FMRI BOLD data

Dataset contains fMRI BOLD data from from five subjects in each of five resting-state conditions. Object is a data frame with the following fields:

"Subject" Subject ID

"Condition" Condition label

"Data" A 150 x 5 matrix containing the standardized BOLD signal from 5 cortical regions.

Author

Corson N. Areshenkoff areshenk@protonmail.com